Refactor FlashMD stepper
