+++ src/flashmd/steppers/__init__.py
@@ -1,5 +1,6 @@
 from .core import AtomisticStepper
 from .symplectic import SymplecticStepper
+from .flashmd import FlashMDStepper
 
 
-__all__ = ["AtomisticStepper", "SymplecticStepper"]
+__all__ = ["AtomisticStepper", "FlashMDStepper", "SymplecticStepper"]
